feat(plugin): migration Dalamud API Level 15
- Bump Dalamud.NET.Sdk 14.0.2 → 15.0.0
- TerritoryChanged et TerritoryType : ushort → uint
- DalamudApiLevel 14 → 15

+++ EorzeaEventsPlugin/LocationDebugSnapshot.cs
@@ -12,7 +12,7 @@ internal sealed class LocationDebugSnapshot
     public string CharacterName { get; init; } = "—";
     public string WorldName { get; init; } = "—";
 
-    public ushort TerritoryId { get; init; }
+    public uint TerritoryId { get; init; }
     public uint MapId { get; init; }
     public string TerritoryName { get; init; } = "—";
